Rename director to snake case
