+++ src/Services.ts
@@ -3,14 +3,14 @@ import ApiClient from '@covid/core/api/ApiClient';
 import OfflineService from '@covid/core/offline/OfflineService';
 import PushNotificationService, {
   PushNotificationApiClient,
-} from '@covid/core/pushNotifications/PushNotificationService';
+} from '@covid/core/push-notifications/PushNotificationService';
 import ContentService from '@covid/core/content/ContentService';
 import { ContentApiClient } from '@covid/core/content/ContentApiClient';
 
 import { AssessmentApiClient } from './core/assessment/AssessmentApiClient';
 import AssessmentService from './core/assessment/AssessmentService';
 import ReduxAssessmentState from './core/assessment/AssessmentState';
-import ExpoPushTokenEnvironment from './core/pushNotifications/expo';
+import ExpoPushTokenEnvironment from './core/push-notifications/expo';
 
 const apiClient = new ApiClient();
 const localStorageService = new LocalStorageService();
